Fix get published
