+++ src/screens/Overview/components/NextMilestone/components/GetPublished/components/Checklist/components/MoreInfo.js
