+++ src/screens/Overview/components/NextMilestone/components/GetPublished/index.js
@@ -1,17 +1,75 @@
 import React from 'react'
-import {getPublishedTitleText} from 'utils/text'
+import {map, uniq, compact, isString} from 'lodash'
+import {getPublishedTitleText, getPublishedDescriptionText} from 'utils/text'
+import {chatInfoUrl, roughDraftInfoUrl, gearSetupInfoUrl} from 'utils/urls'
 import Heading from 'components/Heading'
-import GetPublishedSteps from './components/GetPublishedSteps'
-import GetPublishedFaq from './components/GetPublishedFaq'
+import isStepComplete from './utils/isStepComplete'
+import Checklist from './components/Checklist'
+import DescriptionBlock from './components/DescriptionBlock'
 
 export default ({
   instructor,
-}) => (
-  <div>
-    <Heading level='3'>
-      {getPublishedTitleText}
-    </Heading>
-    <GetPublishedSteps instructor={instructor} />
-    <GetPublishedFaq />
-  </div>
-)
+  firstPageOfLessons,
+}) => {
+
+  const instructorLessonStates = compact(uniq(map(firstPageOfLessons, 'state')))
+
+  const steps = [
+    {
+      isComplete: true,
+      description: 'Create instructor account',
+    },
+    {
+      isComplete: isString(instructor.slack_id),
+      description: 'Join egghead Slack',
+      moreInfoUrl: chatInfoUrl,
+    },
+    {
+      isComplete: isStepComplete(instructorLessonStates, 'claimed'),
+      description: 'Claim new lesson',
+      action: '/lessons/new',
+    },
+    {
+      isComplete: isStepComplete(instructorLessonStates, 'submitted'),
+      description: 'Submit rough draft',
+      moreInfoUrl: roughDraftInfoUrl,
+    },
+    {
+      isComplete: isString(instructor.gear_tracking_id),
+      description: 'Get gear',
+      moreInfoUrl: gearSetupInfoUrl,
+    },
+    {
+      isComplete: isStepComplete(instructorLessonStates, 'updated'),
+      description: 'Re-record with gear',
+      moreInfoUrl: roughDraftInfoUrl,
+    },
+    {
+      isComplete: isStepComplete(instructorLessonStates, 'approved'),
+      description: 'Iterate until approved',
+      moreInfoUrl: roughDraftInfoUrl,
+    },
+    {
+      isComplete: isStepComplete(instructorLessonStates, 'published'),
+      description: 'Publish lesson',
+    },
+  ]
+
+  return (
+    <div>
+      <Heading level='2'>
+        {getPublishedTitleText}
+      </Heading>
+      <DescriptionBlock>
+        {getPublishedDescriptionText}
+      </DescriptionBlock>
+
+      <div className='mt3'>
+        <Checklist 
+          items={steps} 
+          instructorId={instructor.id}
+        />
+      </div>
+    </div>
+  )
+}
