+++ src/screens/Overview/components/NextMilestone/components/GetPublished/utils/isStepComplete/index.test.js
